requirements.txt
aws-cdk-lib==2.205.0
constructs>=10.0.0,<11.0.0
cdklabs.generative_ai_cdk_constructs>=0.1.309
cdk-ecr-deployment>=4.0.0
strands-agents>=0.1.0
python-dotenv>=1.0.0
boto3>=1.34.0
bedrock-agentcore>=0.1.0
streamlit>=1.30.0
